requirements.txt

Flask==3.0.3
Werkzeug==3.0.3


python-dotenv==1.0.1


requests==2.32.3
urllib3==2.2.3


cryptography==42.0.8


reportlab==4.0.9


ebooklib==0.18.0
beautifulsoup4==4.12.3


gunicorn==22.0.0


celery==5.3.4
kombu==5.3.4


psycopg2-binary==2.9.9
SQLAlchemy==2.0.23


redis==5.0.1
flask-session==0.6.0


flask-limiter==3.5.0
flask-wtf==1.2.1


tiktoken==0.5.2

google-cloud-vision>=3.4.0
azure-cognitiveservices-vision-computervision>=0.9.0

opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0

google-generativeai>=0.3.0

iopaint>=1.0.0
